added XAuth code to the project and fought with all the new conflicts. now I just need to get all this stuff working
git-svn-id: http://literalshore.com/svn/kickball/trunk@652 73bdbd85-dd75-0410-9d14-8527f93b1508

+++ Classes/MGTwitterEngine.m
@@ -9,7 +9,7 @@
 #import "MGTwitterEngine.h"
 #import "MGTwitterHTTPURLConnection.h"
 
-#import "NSData+Base64.h"
+#import "NSData+MGTwitterBase64.h"
 
 #define USE_LIBXML 1
 
@@ -1441,7 +1441,7 @@
 #pragma mark -
 
 
-- (NSString *)markUpdate:(unsigned long long)updateID asFavorite:(BOOL)flag
+- (NSString *)markUpdate:(unsigned long)updateID asFavorite:(BOOL)flag
 {
 	NSString *path = nil;
 	MGTwitterRequestType requestType;
